# docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "Fluxend is backend as-a-service platform that allows you to build, deploy, and scale applications without managing infrastructure.",
        "title": "Fluxend API",
        "contact": {
            "name": "API Support",
            "url": "http://github.com/fluxend/fluxend",
            "email": "hello@fluxend.app"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/forms/{formUUID}/fields": {
            "get": {
                "description": "Retrieve a list of all fields in a specific form",
                "tags": [
                    "Form Fields"
                ],
                "summary": "List fields",
                "parameters": [
                    {
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Form UUID",
                        "name": "formUUID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of fields",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "allOf": [
                                            {
                                                "$ref": "#/components/schemas/response.Response"
                                            },
                                            {
                                                "type": "object",
                                                "properties": {
                                                    "content": {
                                                        "type": "array",
                                                        "items": {
                                                            "$ref": "#/components/schemas/form.FieldResponseApi"
                                                        }
                                                    }
                                                }
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "description": "Add a new field to a form",
                "tags": [
                    "Form Fields"
                ],
                "summary": "Create field",
                "parameters": [
                    {
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Form UUID",
                        "name": "formUUID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/form.CreateFormFieldsRequest"
                            }
                        }
                    },
                    "description": "Field details",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Field created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/response.Response"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "content": {
                                                    "$ref": "#/components/schemas/form.FieldResponseApi"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/forms/{formUUID}/fields/{fieldUUID}": {
            "get": {
                "description": "Get details of a specific field",
                "tags": [
                    "Form Fields"
                ],
                "summary": "Retrieve field",
                "parameters": [
                    {
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Form UUID",
                        "name": "formUUID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Field UUID",
                        "name": "fieldUUID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Field details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/response.Response"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "content": {
                                                    "$ref": "#/components/schemas/form.FieldResponseApi"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "put": {
                "description": "Update the details of an existing field",
                "tags": [
                    "Form Fields"
                ],
                "summary": "Update field",
                "parameters": [
                    {
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Form UUID",
                        "name": "formUUID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Field UUID",
                        "name": "fieldUUID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/form.UpdateFormFieldRequest"
                            }
                        }
                    },
                    "description": "Field details",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Field updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/response.Response"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "content": {
                                                    "$ref": "#/components/schemas/form.FieldResponseApi"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "description": "Remove a specific field from the form",
                "tags": [
                    "Form Fields"
                ],
                "summary": "Delete field",
                "parameters": [
                    {
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Form UUID",
                        "name": "formUUID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Field UUID",
                        "name": "fieldUUID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Field deleted"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "//api.fluxend.app"
        }
    ],
    "components": {
...
        "schemas": {
...
            "form.CreateFormFieldsRequest": {
                "type": "object",
                "properties": {
                    "fields": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/form.FieldRequest"
                        }
                    },
                    "projectUUID": {
                        "type": "string"
                    }
                }
            },
...
            "form.FieldRequest": {
                "type": "object",
                "properties": {
                    "date_format": {
                        "description": "fails if provided and field value doesn't match",
                        "type": "string"
                    },
                    "default_value": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "end_date": {
                        "type": "string"
                    },
                    "is_required": {
                        "type": "boolean"
                    },
                    "label": {
                        "description": "required fields",
                        "type": "string"
                    },
                    "max_length": {
                        "type": "integer"
                    },
                    "max_value": {
                        "type": "integer"
                    },
                    "min_length": {
                        "description": "all fields from this point are optional",
                        "type": "integer"
                    },
                    "min_value": {
                        "description": "only applicable for number types",
                        "type": "integer"
                    },
                    "options": {
                        "description": "Options for select/radio types",
                        "type": "string"
                    },
                    "pattern": {
                        "type": "string"
                    },
                    "start_date": {
                        "description": "only applicable for date types",
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    }
                }
            },
            "form.FieldResponseApi": {
                "type": "object",
                "properties": {
                    "createdAt": {
                        "type": "string"
                    },
                    "dateFormat": {
                        "type": "string"
                    },
                    "defaultValue": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "endDate": {
                        "type": "string"
                    },
                    "formUuid": {
                        "type": "string"
                    },
                    "isRequired": {
                        "type": "boolean"
                    },
                    "label": {
                        "type": "string"
                    },
                    "maxLength": {
                        "type": "integer"
                    },
                    "maxValue": {
                        "type": "integer"
                    },
                    "minLength": {
                        "type": "integer"
                    },
                    "minValue": {
                        "type": "integer"
                    },
                    "options": {
                        "type": "string"
                    },
                    "pattern": {
                        "type": "string"
                    },
                    "startDate": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "type": "string"
                    },
                    "uuid": {
                        "type": "string"
                    }
                }
            },
...
            "form.UpdateFormFieldRequest": {
                "type": "object",
                "properties": {
                    "date_format": {
                        "description": "fails if provided and field value doesn't match",
                        "type": "string"
                    },
                    "default_value": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "end_date": {
                        "type": "string"
                    },
                    "is_required": {
                        "type": "boolean"
                    },
                    "label": {
                        "description": "required fields",
                        "type": "string"
                    },
                    "max_length": {
                        "type": "integer"
                    },
                    "max_value": {
                        "type": "integer"
                    },
                    "min_length": {
                        "description": "all fields from this point are optional",
                        "type": "integer"
                    },
                    "min_value": {
                        "description": "only applicable for number types",
                        "type": "integer"
                    },
                    "options": {
                        "description": "Options for select/radio types",
                        "type": "string"
                    },
                    "pattern": {
                        "type": "string"
                    },
                    "projectUUID": {
                        "type": "string"
                    },
                    "start_date": {
                        "description": "only applicable for date types",
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    }
                }
            },
...
            "response.Response": {
                "type": "object",
                "properties": {
                    "content": {},
                    "errors": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "success": {
                        "type": "boolean"
                    }
                }
            },
...
        }
    }
}